Compile the full engine on Windows/MSVC
Extends the M11.d.5 hosted-runner and physical-hardware D3D12 reach to
the full `cmake --build --preset dev` on Windows/MSVC. 746 targets build
green from a clean tree; `ctest --test-dir build/dev -R rhi_d3d12` and
`render.golden_backends` remain matched, and 84% of the full ctest suite
passes on a first Windows run. The remainder is largely downstream from
the CPU-time-clock gap `tests/harness/src/budget.cpp` already names —
"the platform that cannot keeps the behaviour it has always had, named
rather than silently different" — and is not addressed here.

Portability fixes, by pattern:

POSIX APIs -> C++17/20 standard equivalents
  - tools/build/src/patch.cpp:            _exit()      -> std::_Exit
  - tools/build/src/artefact_store.cpp:   chmod()      -> std::filesystem::permissions
  - tools/build/src/main.cpp:             getcwd()     -> std::filesystem::current_path
  - tools/build/tests/test_build_service.cpp:                                        same
  - src/core/diagnostics/tests/test_reproduction.cpp: mkdtemp() + /tmp
    -> std::filesystem::temp_directory_path
  - samples/07-fidelity/frame.cpp: __builtin_popcount -> std::popcount

POSIX-only tests guarded at CMake, not portablized
  - tools/build/tests/CMakeLists.txt: build_patch (fork + waitpid, no Win32 equivalent
    without a full rewrite; the soft-interrupt path still covers rollback on every host)
  - src/save/tests/CMakeLists.txt:    save_kill_nine (fork + SIGKILL + waitpid; the
    in-process half in test_archive.cpp still runs on Windows)

POSIX / Win32 forked APIs
  - tools/material/tests/test_material_cook.cpp: popen/pclose -> _popen/_pclose on _WIN32
  - src/vfx/tests/test_vfx_runtime.cpp: clock_gettime(CLOCK_THREAD_CPUTIME_ID)
    guarded to match the pattern already used in five other test files; falls back to
    steady_clock on Windows where no per-thread CPU clock is standard

MSVC compiler quirks
  - src/core/memory/include/cy/core/memory/hash_map.h: [[no_unique_address]] does not
    fold on the MSVC ABI; add [[msvc::no_unique_address]] so
    sizeof(HashMap<K, Unit>::Entry) == sizeof(K), which HashSet's own taxonomy asserts.
    GCC and Clang ignore unknown attribute namespaces per [dcl.attr.grammar]/6.
  - src/servers/audio/include/cy/servers/audio/server.h: AudioServer::VoiceSlot has a
    std::atomic<u32> member which deletes copy/move. Array<VoiceSlot>::reallocate_to
    still instantiates move-construction under MSVC's stricter template model even in
    the if-constexpr-untaken branch, so give VoiceSlot an explicit move-ctor that
    loads the atomic. Safe: relocation runs during initialize(), before the audio
    thread starts.
  - src/rendering/skinning/shaders/embed_msl.py + skin_msl.h,
    samples/10-world/shaders/embed_msl.py + world_visual_msl.h: MSVC's per-token
    logical-source-line limit rejects raw string literals that span more than a few
    thousand lines. Regenerate both MSL headers as adjacent chunked raw literals
    (~300 lines per chunk) which the compiler concatenates. GCC / Clang treat the
    concatenated form identically to one literal. Both generators updated so future
    regenerations produce the same shape.
  - src/core/determinism/tests/test_profile.cpp,
    src/core/serialize/tests/test_text.cpp: 0.0F / 0.0F and 1.0F / 0.0F are refused
    at compile time by MSVC's constant folder under /permissive-. Use
    std::numeric_limits<f32>::quiet_NaN() / ::infinity() — same intent, portable.
  - tests/harness/include/cy/test/test.h: MSVC's <string_view> declares an operator<<
    for basic_ostream inline, referencing basic_ostream::iostate. Include <ostream>
    from the one file every test transitively includes; GCC / Clang are unaffected.

/WX warnings
  - samples/07-fidelity/shade.cpp: static_cast<u32> on width-modulo results (C4244)
  - samples/07-fidelity/frame.cpp: adds <bit> alongside the popcount switch
  - src/graph/tests/test_lowering.cpp: static_cast<u32> on Span::size() (C4267)

Dependency configuration
  - cmake/dependencies.cmake: force Jolt onto the dynamic CRT (/MD) on WIN32. Jolt
    defaults USE_STATIC_MSVC_RUNTIME_LIBRARY=ON, which fails to link against the
    rest of the tree (LNK2038 across every Jolt object).

None of these change POSIX behaviour. Every #if defined(_WIN32) or NOT WIN32 branch
guards the Windows path only; the Linux and macOS paths are the exact code that was
already shipping.

+++ samples/07-fidelity/shade.cpp
@@ -511,8 +511,8 @@ private:
             ++report.unreconstructed;
             continue;
         }
-        const u32 column = pixel % capture.width;
-        const u32 row = pixel / capture.width;
+        const u32 column = static_cast<u32>(pixel % capture.width);
+        const u32 row = static_cast<u32>(pixel / capture.width);
         const Vec2 centre{static_cast<f32>(column) + 0.5F, static_cast<f32>(row) + 0.5F};
         Expected<rendering::vg::SurfaceAttributes, Error> attributes =
             rendering::vg::reconstruct_surface(
